% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_response}
\alias{get_response}
\title{The engine of the request functions}
\usage{
get_response(dota_id, dota_api_method, dota_api_category, api_version, args)
}
\arguments{
\item{dota_id}{Can take one of three values.

\itemize{
 \item 570 - Dota 2 (default)
 \item 816 - Dota 2 Internal Test
 \item 205790 - Dota 2 Beta Test
}}

\item{dota_api_method}{The api method.}

\item{dota_api_category}{One of IDOTA2Match or IEconDOTA2.}

\item{api_version}{The api version.}

\item{args}{The query arguments for GET.}
}
\value{
A response to be used in the rest of the get_* family functions.

%makes documentation available but removes function from index
}
\description{
The function used to fetch the response content inside the the API functions.
}
\details{
This function is of no particular use to the user, but might be useful for anyone who
would like to dive deeper into the package development. This function is used to GET the
response's content from Steam's API. The function is the core function of the rest of
the get_* family functions.
}
\keyword{internal}

