% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method_dota_api_class.R
\name{print.dota_api}
\alias{print.dota_api}
\title{print method for dota_api class}
\usage{
\method{print}{dota_api}(x)
}
\arguments{
\item{x}{A dota_api object.}
}
\value{
Prints the content element of the dota_api list.
}
\description{
print method for dota_api class
}
\details{
A method to print a dota_api object. Only the content will be printed on screen for a dota_api
object because this is of interest to the user. The dota_api object will usually be a list of
three elements containing the request url and the response apart from the content.
}

