% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_response}
\alias{get_response}
\title{The engine of the request functions}
\usage{
get_response(dota_id, dota_api_method, dota_api_category, api_version, ...)
}
\arguments{
\item{dota_id}{Can take three values.

\itemize {

}570 for Dota 2, 816 for Dota 2 Internal Test or
205790 for Dota 2 beta test.}

\item{dota_api_method}{The api method.}

\item{dota_api_category}{One of IDOTA2Match or IEconDOTA2.}

\item{api_version}{The api version.}

\item{...}{The query arguments for GET.}
}
\value{
A response to be used in the rest of the get_* family functions.
}
\description{
The function used to fetch the response content inside the the API functions.
}
\details{
This function is of no particular use to the user, but might be useful for anyone who
would like to dive deeper into the package development. This function is used to GET the
response's content from Steam's API. The function is the core function of the rest of
the get_* family functions.
}
\keyword{internal}

